Add files via upload
Blueprint created for a Zigbee wall switch with 3 buttons which are acting switches in HA.
I want to achieve that switch 2 and 3 will act as push buttons in an interleave lock for controlling a Shelly 2PM cover. Basically, button 2 and 3 from the wall switch will control the opening/closing of the Shelly 2PM cover device.